move multiply.hpp to /prim use test utility for tests of scalar * mat
