+++ stan/math/opencl/prim/categorical_logit_glm_lpmf.hpp
@@ -6,7 +6,7 @@
 #include <stan/math/opencl/rev/operands_and_partials.hpp>
 #include <stan/math/opencl/matrix_cl.hpp>
 #include <stan/math/opencl/copy.hpp>
-#include <stan/math/opencl/multiply.hpp>
+#include <stan/math/opencl/prim/multiply.hpp>
 #include <stan/math/opencl/kernel_generator.hpp>
 #include <stan/math/opencl/kernels/categorical_logit_glm_lpmf.hpp>
