+++ stan/math/opencl/prim/normal_id_glm_lpdf.hpp
@@ -6,7 +6,7 @@
 #include <stan/math/opencl/rev/operands_and_partials.hpp>
 #include <stan/math/opencl/copy.hpp>
 #include <stan/math/opencl/matrix_cl.hpp>
-#include <stan/math/opencl/multiply.hpp>
+#include <stan/math/opencl/prim/multiply.hpp>
 #include <stan/math/opencl/kernel_generator.hpp>
 #include <stan/math/opencl/prim/size.hpp>
 #include <stan/math/prim/meta.hpp>
