+++ stan/math/opencl/prim/ordered_logistic_glm_lpmf.hpp
@@ -7,7 +7,7 @@
 #include <stan/math/opencl/copy.hpp>
 #include <stan/math/opencl/kernel_generator.hpp>
 #include <stan/math/opencl/matrix_cl.hpp>
-#include <stan/math/opencl/multiply.hpp>
+#include <stan/math/opencl/prim/multiply.hpp>
 #include <stan/math/opencl/kernels/ordered_logistic_glm_lpmf.hpp>
 
 #include <stan/math/prim/meta.hpp>
