+++ test/unit/math/opencl/kernel_generator/binary_operation_test.cpp
@@ -3,7 +3,7 @@
 #include <stan/math/opencl/kernel_generator.hpp>
 #include <stan/math/opencl/matrix_cl.hpp>
 #include <stan/math/opencl/copy.hpp>
-#include <stan/math/opencl/multiply.hpp>
+#include <stan/math/opencl/prim/multiply.hpp>
 #include <test/unit/math/opencl/kernel_generator/reference_kernel.hpp>
 #include <test/unit/util.hpp>
 #include <Eigen/Dense>
